+++ docs/src/demos/Examples/Book/index.vue
@@ -70,7 +70,8 @@
 </template>
 
 <script>
-import { Editor, EditorContent, defaultExtensions } from '@tiptap/vue-starter-kit'
+import { Editor, EditorContent } from '@tiptap/vue-2'
+import { defaultExtensions } from '@tiptap/starter-kit'
 import { content } from './content.js'
 
 export default {