+++ docs/src/demos/Examples/Formatting/index.vue
@@ -43,7 +43,8 @@
 </template>
 
 <script>
-import { defaultExtensions, Editor, EditorContent } from '@tiptap/vue-starter-kit'
+import { Editor, EditorContent } from '@tiptap/vue-2'
+import { defaultExtensions } from '@tiptap/starter-kit'
 import TextAlign from '@tiptap/extension-text-align'
 import Highlight from '@tiptap/extension-highlight'