+++ docs/src/tests/core/commands/setContent.spec.js
@@ -1,6 +1,6 @@
 context('setContent', () => {
   before(() => {
-    cy.visit('/demos/Examples/default')
+    cy.visit('/demos/Examples/Default/Vue')
   })
 
   it('returns true for the setContent command', () => {