+++ packages/vue-2/src/index.ts
@@ -1,0 +1,4 @@
+export * from '@tiptap/core'
+export * from './VueRenderer'
+export * from './VueNodeViewRenderer'
+export * from './EditorContent'