+++ packages/vue-3/src/useEditor.ts
@@ -1,0 +1,17 @@
+import { onMounted, onBeforeUnmount, ref } from 'vue'
+import { EditorOptions } from '@tiptap/core'
+import { Editor } from './Editor'
+
+export const useEditor = (options: Partial<EditorOptions> = {}) => {
+  const editor = ref()
+
+  onMounted(() => {
+    editor.value = new Editor(options)
+  })
+
+  onBeforeUnmount(() => {
+    editor.value.destroy()
+  })
+
+  return editor
+}