+++ packages/extension-gapcursor/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.12](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-gapcursor@2.0.0-alpha.11...@tiptap/extension-gapcursor@2.0.0-alpha.12) (2021-02-26)
+
+**Note:** Version bump only for package @tiptap/extension-gapcursor
+
+
+
+
+
 # [2.0.0-alpha.11](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/extension-gapcursor@2.0.0-alpha.10...@tiptap/extension-gapcursor@2.0.0-alpha.11) (2021-02-16)
 
 **Note:** Version bump only for package @tiptap/extension-gapcursor